refactor(attachments): extract staging out of Wizard, add a shared backend knob
Async interactions need what wizards already had: a way to hold an
attachment as a plain string while the work is in flight, because
neither has a model to hang it off until the work lands. None of that
machinery was wizard-specific — staging to a backend cache, reviving a
token through whichever backend minted it, and the uniform view over an
AS Blob vs a Shrine UploadedFile are all backend plumbing.

Plutonium::Attachments now holds it. Wizard::Attachments keeps its
published surface — resolve, stage_upload, validation_errors, field?,
attachment_backend and the Resolved constant — as a thin layer whose one
remaining job is defaulting the backend to the wizard's own setting.
field? stays there too: it reads a step input's options, a shape nothing
outside the wizard DSL has.

Backend config is now layered rather than per-subsystem. The new
config.attachment_backend is the shared default; config.wizards
.attachment_backend narrows it to wizards as before, and the new
config.async_interactions.attachment_backend narrows it to runs. An app
setting only the shared one covers both.

No behaviour change: the wizard attachment suites pass untouched.
