+++ lib/plutonium/interaction/async/configuration.rb
@@ -17,6 +17,16 @@ module Plutonium
         #   progress write before ReapJob considers it stalled
         attr_accessor :stall_after
 
+        # @return [Symbol, nil] the storage backend used to stage an attachment
+        #   attribute into the run's options — `:active_storage` or `:shrine`.
+        #   `nil` falls back to `config.attachment_backend`, then to auto-detection.
+        #
+        #   A file cannot ride the options column: it is JSON, and the request's
+        #   tempfile is gone by the time the job runs. So an attachment attribute is
+        #   uploaded to the backend's cache at dispatch and carried as its token.
+        #   This is the knob for which backend does that, for runs alone.
+        attr_accessor :attachment_backend
+
         def initialize
           @enabled = false
           @queue = :default
